Add files via upload
Index will be our main page. Currently it's set to grab data from test.php but will eventually be our homepage for all users, logged in or out.
test.php has a bunch of test functions on there in PHP and JavaScript to help everyone get familiar with the kinds of AJAX coding we may decide to use to update our page.